% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodnessOfFit.R
\name{cn.test}
\alias{cn.test}
\title{Central normality test}
\usage{
cn.test(x, transformer = NULL, robust = FALSE)
}
\arguments{
\item{x}{vector of input data, of at least length 5.}

\item{transformer}{A transformer object created using
\code{find_transformation_parameters}. Optional, if present residuals are
determined from x after transformation.}

\item{robust}{Determines the use of the strict test for central normality
(\code{FALSE}) or the more robust version where test statistics where determined
from normally distributed data with 10\% outliers (\code{TRUE}).}
}
\value{
list with mean absolute error (\code{tau}), critical value (at
significance level = 0.95) of the test statistic (\code{tau_critical}) and
p-value (\code{p_value}) for (empirical) central normality test.
}
\description{
Assesses central normality of input data.
}
