% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationSkeleton.R
\name{create_transformer_skeleton}
\alias{create_transformer_skeleton}
\title{Create transformation object skeleton}
\usage{
create_transformer_skeleton(method, lambda = 1, shift = 0, scale = 1)
}
\arguments{
\item{method}{Transformation method. Can be \code{"none"}, \code{"box_cox"} or
\code{"yeo_johnson"}.}

\item{lambda}{Value of the transformation parameter lambda. Can also be
changed using the \code{set_lambda} method.}

\item{shift}{Value of the shift parameter. Can also be changed using the
\code{set_shift} method.}

\item{scale}{Value of the scale parameter. Can also be changed using the
\code{set_scale} method.}
}
\value{
A transformer object
}
\description{
Creates skeleton objects. This generates objects without fitting
parameters. This is primarily intended for creating transformers
externally, where fitting parameters are known.
}
