% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodnessOfFit.R
\name{ecn.test}
\alias{ecn.test}
\title{Empirical central normality test}
\usage{
ecn.test(x, transformer = NULL)
}
\arguments{
\item{x}{vector of input data, of at least length 5.}

\item{transformer}{A transformer object created using
\code{find_transformation_parameters}. Optional, if present residuals are
determined from x after transformation.}
}
\value{
list with mean absolute error (\code{tau}), critical value (at
significance level = 0.95) of the test statistic (\code{tau_critical}) and
p-value (\code{p_value}) for the empirical central normality test.
}
\description{
Assesses central normality of input data using an empirical test.
}
