% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodnessOfFit.R
\name{ecn_test}
\alias{ecn_test}
\title{Empirical central normality test}
\usage{
ecn_test(
  x = NULL,
  transformer = NULL,
  tau = NULL,
  n = NULL,
  outlier_rate = 0.1,
  kappa = 0.8
)
}
\arguments{
\item{x}{vector of input data, of at least length 5.}

\item{transformer}{A transformer object created using
\code{find_transformation_parameters}. Optional, if present residuals are
determined from x after transformation.}
}
\value{
list with mean absolute error (\code{tau}), critical value (at
significance level = 0.95) of the test statistic (\code{tau_critical}) and
p-value (\code{p_value}) for the empirical central normality test.
}
\description{
Assesses central normality of input data using an empirical test.
}
