% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{huber_estimate}
\alias{huber_estimate}
\title{Huber M-estimate}
\usage{
huber_estimate(x, k = 1.28, tol = 1e-04)
}
\arguments{
\item{x}{Vector of numeric values for which the location and scale should be
estimated.}

\item{k}{Numeric value > 0 that the determines the value beyond which the
signal is winsorized.}

\item{tol}{Tolerance for the iterative procedure.}
}
\value{
list with location estimate \code{"mu"} and scale estimate \code{"sigma"}.
}
\description{
Iteratively computes M-estimates for location and scale. These
are robust estimates of the mean and standard deviation of the data.
}
