% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsMutatorsLambda.R
\name{get_lambda}
\alias{get_lambda}
\alias{get_lambda,transformationPowerTransform-method}
\alias{get_lambda,transformationBoxCox-method}
\alias{get_lambda,transformationYeoJohnson-method}
\title{Get lambda value}
\usage{
get_lambda(object, ...)

\S4method{get_lambda}{transformationPowerTransform}(object, ...)

\S4method{get_lambda}{transformationBoxCox}(object, ...)

\S4method{get_lambda}{transformationYeoJohnson}(object, ...)
}
\arguments{
\item{object}{Transformer object}

\item{...}{Unused arguments}
}
\value{
Lambda value of the transformer.
}
\description{
Get the lambda value of a transformer object.
}
