% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsMutatorsLambda.R
\name{set_lambda}
\alias{set_lambda}
\alias{set_lambda,transformationPowerTransform-method}
\alias{set_lambda,transformationBoxCox-method}
\alias{set_lambda,transformationYeoJohnson-method}
\title{Set lambda value}
\usage{
set_lambda(object, lambda, ...)

\S4method{set_lambda}{transformationPowerTransform}(object, lambda, ...)

\S4method{set_lambda}{transformationBoxCox}(object, lambda, ...)

\S4method{set_lambda}{transformationYeoJohnson}(object, lambda, ...)
}
\arguments{
\item{object}{Transformer object}

\item{lambda}{Lambda value}

\item{...}{Unused arguments}
}
\value{
Transformer object with updated lambda value.
}
\description{
Set the lambda value of a transformer object.
}
