% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.transform-package.R
\docType{package}
\name{power.transform}
\alias{power.transform-package}
\alias{power.transform}
\title{power.transform: Transform Data to Normality using Power Transformations}
\description{
This package was originally based on, and contains code from, the familiar
package (\url{https://cran.r-project.org/package=familiar}), under the EUPL
license.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oncoray/power.transform}
  \item Report bugs at \url{https://github.com/oncoray/power.transform/issues}
}

}
\author{
\strong{Maintainer}: Alex Zwanenburg \email{alex.zwanenburg@outlook.com} (\href{https://orcid.org/0000-0002-0342-9545}{ORCID})

Authors:
\itemize{
  \item Alex Zwanenburg \email{alex.zwanenburg@outlook.com} (\href{https://orcid.org/0000-0002-0342-9545}{ORCID})
  \item Steffen Löck
}

Other contributors:
\itemize{
  \item German Cancer Research Center (DKFZ) [copyright holder]
}

}
