% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsMutatorsScale.R
\name{get_scale}
\alias{get_scale}
\alias{get_scale,transformationPowerTransform-method}
\alias{get_scale,transformationBoxCox-method}
\alias{get_scale,transformationYeoJohnson-method}
\title{Get scale value}
\usage{
get_scale(object, ...)

\S4method{get_scale}{transformationPowerTransform}(object, ...)

\S4method{get_scale}{transformationBoxCox}(object, ...)

\S4method{get_scale}{transformationYeoJohnson}(object, ...)
}
\arguments{
\item{object}{Transformer object}

\item{...}{Unused arguments}
}
\value{
scale value of the transformer.
}
\description{
Get the scale value of a transformer object.
}
