% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsMutatorsScale.R
\name{set_scale}
\alias{set_scale}
\alias{set_scale,transformationPowerTransform-method}
\alias{set_scale,transformationBoxCox-method}
\alias{set_scale,transformationYeoJohnson-method}
\title{Set scale value}
\usage{
set_scale(object, scale, ...)

\S4method{set_scale}{transformationPowerTransform}(object, scale, ...)

\S4method{set_scale}{transformationBoxCox}(object, scale, ...)

\S4method{set_scale}{transformationYeoJohnson}(object, scale, ...)
}
\arguments{
\item{object}{Transformer object}

\item{scale}{scale value}

\item{...}{Unused arguments}
}
\value{
Transformer object with updated scale value.
}
\description{
Set the scale value of a transformer object.
}
