% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsMutatorsShift.R
\name{get_shift}
\alias{get_shift}
\alias{get_shift,transformationPowerTransform-method}
\alias{get_shift,transformationBoxCox-method}
\alias{get_shift,transformationYeoJohnson-method}
\title{Get shift value}
\usage{
get_shift(object, ...)

\S4method{get_shift}{transformationPowerTransform}(object, ...)

\S4method{get_shift}{transformationBoxCox}(object, ...)

\S4method{get_shift}{transformationYeoJohnson}(object, ...)
}
\arguments{
\item{object}{Transformer object}

\item{...}{Unused arguments}
}
\value{
shift value of the transformer.
}
\description{
Get the shift value of a transformer object.
}
