% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsMutatorsShift.R
\name{set_shift}
\alias{set_shift}
\alias{set_shift,transformationPowerTransform-method}
\alias{set_shift,transformationBoxCox-method}
\alias{set_shift,transformationYeoJohnson-method}
\title{Set shift value}
\usage{
set_shift(object, shift, ...)

\S4method{set_shift}{transformationPowerTransform}(object, shift, ...)

\S4method{set_shift}{transformationBoxCox}(object, shift, ...)

\S4method{set_shift}{transformationYeoJohnson}(object, shift, ...)
}
\arguments{
\item{object}{Transformer object}

\item{shift}{Shift value}

\item{...}{Unused arguments}
}
\value{
Transformer object with updated shift value.
}
\description{
Set the shift value of a transformer object.
}
