% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsTransformationMethod.R
\name{get_transformation_method}
\alias{get_transformation_method}
\alias{get_transformation_method,transformationPowerTransform-method}
\title{Get transformation method}
\usage{
get_transformation_method(object, ...)

\S4method{get_transformation_method}{transformationPowerTransform}(object, ...)
}
\arguments{
\item{object}{Transformer object}

\item{...}{Unused arguments}
}
\value{
Transformation method
}
\description{
Get the transformation method of a transformer object.
}
