% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationObjects.R
\docType{class}
\name{transformationNone-class}
\alias{transformationNone-class}
\title{No transformation object}
\description{
This class is for transformers that do not alter the data.
}
\section{Slots}{

\describe{
\item{\code{method}}{Main transformation method, i.e. \code{"none"}.}

\item{\code{complete}}{Indicates whether transformation parameters were set.}
}}

