% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationObjects.R
\docType{class}
\name{transformationPowerTransform-class}
\alias{transformationPowerTransform-class}
\title{Generic transformation object}
\description{
This is the superclass for transformation objects.
}
\section{Slots}{

\describe{
\item{\code{method}}{Main transformation method.}

\item{\code{complete}}{Indicates whether transformation parameters were set.}

\item{\code{version}}{Version of the power.transform package that was used to create
the transformation objecst.}
}}

