% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.transform-package.R
\docType{package}
\name{power.transform}
\alias{power.transform}
\alias{power.transform-package}
\title{power.transform: Power-transformations for Transforming Feature Distributions
to Normality}
\description{
This package is based on, and contains code from, the familiar package
(\url{https://cran.r-project.org/package=familiar}), under the EUPL license.
}
\author{
\strong{Maintainer}: Alex Zwanenburg \email{alexander.zwanenburg@nct-dresden.de} (\href{https://orcid.org/0000-0002-0342-9545}{ORCID})

Other contributors:
\itemize{
  \item German Cancer Research Center (DKFZ) [copyright holder]
}

}
